+++ CHANGES.rst
@@ -10,6 +10,8 @@ RIOS.CORE Change History
   enumeration and enumerationSet).
 * Fixed issue where form event failure text was not being checked for the
   default localization.
+* Added validation logic to ensure tag identifiers don't collide with page or
+  field IDs.
 
 
 0.5.0 (2015-09-25)